Cast switch indices to placate HP's cc.

+++ tests/mpn/t-bdiv.c
@@ -218,7 +218,7 @@ main (int argc, char **argv)
       if (t % 17 == 0)
 	dp[0] = GMP_NUMB_MAX;
 
-      switch (t % 16)
+      switch ((int) t % 16)
 	{
 	case 0:
 	  clearn = random_word (rands) % nn;
